+++ bundle/src/main/java/com/adobe/acs/commons/httpcache/config/impl/keys/helper/KeyValueMapWrapperBuilder.java
@@ -2,7 +2,7 @@
  * #%L
  * ACS AEM Commons Bundle
  * %%
- * Copyright (C) 2016 Adobe
+ * Copyright (C) 2015 Adobe
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,17 @@
  * limitations under the License.
  * #L%
  */
-@Version("5.0.0")
-package com.adobe.acs.commons.redirectmaps.models;
+package com.adobe.acs.commons.httpcache.config.impl.keys.helper;
 
-import org.osgi.annotation.versioning.Version;
+/**
+ * KeyValueMapWrapperBuilder - builds key / values for different key representations
+ * For example: key / values from request cookies, request parameters, value map values
+ */
+public interface KeyValueMapWrapperBuilder {
+
+    /**
+     * Builds a KeyValueMapWrapper that can be used for the cache key
+     * @return
+     */
+    KeyValueMapWrapper build();
+}